+++ src/werkzeug/test.py
@@ -29,6 +29,7 @@ from urllib.request import Request as _UrllibRequest
 from ._internal import _get_environ
 from ._internal import _make_encode_wrapper
 from ._internal import _to_bytes
+from ._internal import _wsgi_decoding_dance
 from ._internal import _wsgi_encoding_dance
 from .datastructures import AnyHeaders
 from .datastructures import Authorization
@@ -453,19 +454,25 @@ class EnvironBuilder:
             self.mimetype = mimetype
 
     @classmethod
-    def from_environ(cls, environ: WSGIEnvironment, **kwargs,) -> "EnvironBuilder":
+    def from_environ(cls, environ: WSGIEnvironment, **kwargs) -> "EnvironBuilder":
         """Turn an environ dict back into a builder. Any extra kwargs
         override the args extracted from the environ.
+
+        .. versionchanged:: 2.0
+            Path and query values are passed through the WSGI decoding
+            dance to avoid double encoding.
 
         .. versionadded:: 0.15
         """
         headers = Headers(EnvironHeaders(environ))
         out = {
-            "path": environ["PATH_INFO"],
+            "path": _wsgi_decoding_dance(environ["PATH_INFO"]),
             "base_url": cls._make_base_url(
-                environ["wsgi.url_scheme"], headers.pop("Host"), environ["SCRIPT_NAME"],
+                environ["wsgi.url_scheme"],
+                headers.pop("Host"),
+                _wsgi_decoding_dance(environ["SCRIPT_NAME"]),
             ),
-            "query_string": environ["QUERY_STRING"],
+            "query_string": _wsgi_decoding_dance(environ["QUERY_STRING"]),
             "method": environ["REQUEST_METHOD"],
             "input_stream": environ["wsgi.input"],
             "content_type": headers.pop("Content-Type", None),
@@ -753,18 +760,17 @@ class EnvironBuilder:
         def _path_encode(x):
             return _wsgi_encoding_dance(url_unquote(x, self.charset), self.charset)
 
-        qs = _wsgi_encoding_dance(self.query_string)
-
+        raw_uri = _wsgi_encoding_dance(self.request_uri, self.charset)
         result.update(
             {
                 "REQUEST_METHOD": self.method,
                 "SCRIPT_NAME": _path_encode(self.script_root),
                 "PATH_INFO": _path_encode(self.path),
-                "QUERY_STRING": qs,
+                "QUERY_STRING": _wsgi_encoding_dance(self.query_string, self.charset),
                 # Non-standard, added by mod_wsgi, uWSGI
-                "REQUEST_URI": _wsgi_encoding_dance(self.request_uri),
+                "REQUEST_URI": raw_uri,
                 # Non-standard, added by gunicorn
-                "RAW_URI": _wsgi_encoding_dance(self.request_uri),
+                "RAW_URI": raw_uri,
                 "SERVER_NAME": self.server_name,
                 "SERVER_PORT": str(self.server_port),
                 "HTTP_HOST": self.host,