+++ CHANGES.rst
@@ -77,6 +77,8 @@ Unreleased
     add an ``Authorization`` header. It can be an ``Authorization``
     object or a ``(username, password)`` tuple for ``Basic`` auth.
     :pr:`1809`
+-   ``EnvironBuilder.from_environ`` decodes values encoded for WSGI, to
+    avoid double encoding the new values. :pr:`1959`
 -   The default stat reloader will watch Python files under
     non-system/virtualenv ``sys.path`` entries, which should contain
     most user code. It will also watch all Python files under